pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
# Check https://flit.readthedocs.io/en/latest/pyproject_toml.html for all available sections
name = "ansys-sound-core"
version = "0.1.dev0"
description = "A Python wrapper for Ansys DPF Sound."
readme = "README.rst"
requires-python = ">=3.9,<4"
license = {file = "LICENSE"}
authors = [
    {name = "ANSYS, Inc.", email = "pyansys.core@ansys.com"},
]
maintainers = [
    {name = "ANSYS, Inc.", email = "pyansys.core@ansys.com"},
]
classifiers=[
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Windows",
]
dependencies = [
    "ansys-dpf-core>=0.9.0,<1",
    "matplotlib>=3.8.2,<4",
]

[project.optional-dependencies]
tests = [
    "ansys-dpf-core==0.12.1",
    "matplotlib==3.8.4",
    "pytest==8.2.1",
    "pytest-cov==5.0.0",
    "pytest-dependency==0.6.0",
    "platformdirs==4.2.0"
]
doc = [
    "Sphinx==7.3.7",
    "ansys-sphinx-theme==0.16.4",
    "numpydoc==1.7.0",
    "sphinx-autodoc-typehints==2.1.0",
    "sphinx-copybutton==0.5.2",
    "sphinx-gallery==0.15.0",
    "sphinx-design==0.5.0",
    "vtk==9.3.0",
    "trame==3.6.1",
    "trame_vtk==2.8.8",
    "pyvista==0.43.8",
    "pypandoc==1.12.0",
    "jupyter-sphinx==0.5.3",
    "sphinx-autobuild==2024.4.16",
    "platformdirs==4.2.0"
]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
force_sort_within_sections = true
line_length = 100
default_section = "THIRDPARTY"
src_paths = ["doc", "src", "tests"]

[tool.coverage.run]
source = ["ansys.sound.core"]

[tool.coverage.report]
show_missing = true

[tool.pytest.ini_options]
minversion = "7.1"
addopts = "-ra --cov=ansys.sound.core --cov-report html:.cov/html --cov-report xml:.cov/xml --cov-report term -vv"
testpaths = ["tests"]
